added helper for enable debug mode , removed log checker
