+++ .github/workflows/product-creation-tests.yml
@@ -437,7 +437,6 @@ jobs:
           sudo apt-get install -y xvfb
 
       - name: Run CAPI Pixel E2E tests
-        if: false
         env:
           TEST_PRODUCT_ID: ${{ env.TEST_PRODUCT_ID }}
         run: |
@@ -469,7 +468,6 @@ jobs:
 
       - name: Run E2E tests
         continue-on-error: true
-        if: false
         env:
           WP_USERNAME: admin
           WP_PASSWORD: admin
