+++ tests/e2e/test-helpers.js
@@ -823,6 +823,44 @@ async function deleteFeedFile(filePath) {
   }
 }
 
+// Helper function to ensure debug mode is enabled
+async function ensureDebugModeEnabled(page) {
+  console.log('🔍 Checking debug mode status...');
+
+  // Navigate to Facebook settings page
+  await page.goto(`${process.env.WORDPRESS_URL}/wp-admin/admin.php?page=wc-facebook`, {
+    waitUntil: 'domcontentloaded',
+    timeout: TIMEOUTS.EXTRA_LONG
+  });
+
+  // Click Troubleshooting toggle to expand drawer
+  const troubleshootingToggle = page.locator('#toggle-troubleshooting-drawer');
+  await troubleshootingToggle.waitFor({ state: 'visible', timeout: TIMEOUTS.LONG });
+  await troubleshootingToggle.click();
+  await page.waitForTimeout(TIMEOUTS.INSTANT);
+
+  // Check debug mode checkbox status
+  const debugModeCheckbox = page.locator('#wc_facebook_enable_debug_mode');
+  await debugModeCheckbox.waitFor({ state: 'visible', timeout: TIMEOUTS.MEDIUM });
+
+  const isChecked = await debugModeCheckbox.isChecked();
+
+  if (!isChecked) {
+    console.log('⚙️ Enabling debug mode...');
+    await debugModeCheckbox.check();
+
+    // Save changes
+    const saveButton = page.locator('input[name="save_shops_settings"]');
+    await saveButton.click();
+    await page.waitForLoadState('domcontentloaded');
+    console.log('✅ Debug mode enabled');
+    return false; // Was not enabled, now is
+  } else {
+    console.log('✅ Debug mode already enabled');
+    return true; // Was already enabled
+  }
+}
+
 module.exports = {
   baseURL,
   username,
@@ -849,5 +887,6 @@ module.exports = {
   setProductTitle,
   exactSearchSelect2Container,
   generateProductFeedCSV,
-  deleteFeedFile
+  deleteFeedFile,
+  ensureDebugModeEnabled
 };
